+++ DmmTimetabling/src/com/timetabling/server/data/entities/timetabling/tt/TeacherTT.java
@@ -40,9 +40,9 @@ public class TeacherTT implements Markable {
 		return lessons;
 	}
 	
-	public double getMarkForVersion(Version version) {
+	public float getMarkForVersion(Version version) {
 		// TODO
-		return 1d;
+		return 1f;
 	}
 
 }
