+++ DmmTimetabling/src/com/timetabling/server/generating/Markable.java
@@ -4,6 +4,6 @@ import com.timetabling.server.data.entities.timetabling.Version;
 
 public interface Markable {
 
-	double getMarkForVersion(Version version);
+	float getMarkForVersion(Version version);
 	
 }
